requirements.txt
asgiref==3.6.0
autopep8==2.0.2
build==0.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.2
django-ipware==5.0.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-recaptcha==2.2.3
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pip-tools==6.13.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyproject_hooks==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
